Don't clobber the following Symbol instance methods when using ruby 1.9: [], <, <=, >, >=
This sizeable commit is necessary due to the fact that Sequel was
violating one of it's own principles, which is that it does not
override methods defined by ruby, it only adds methods not already
defined.  Ruby 1.9 adds Symbol#[] and includes Comparable in Symbol,
which define <, <=, >, and >=.

If you are using 1.9, code like this will stop working:

  dataset.filter(:number > 1)
  dataset.filter(:number >= 2)
  dataset.filter(:name < 'M')
  dataset.filter(:name <= 'I')
  dataset.filter(:is_bool[:x])

This obviously breaks backwards compatibility for 1.9 users, but
it's necessary as we can't allow Sequel to completely change the
meaning of existing methods on core classes.  Thankfully, the fix
is fairly easy, if a little verbose:

  dataset.filter(:number.sql_number > 1)
  dataset.filter(:number.sql_number >= 2)
  dataset.filter(:name.sql_string < 'M')
  dataset.filter(:name.sql_string <= 'I')
  dataset.filter(:is_bool.sql_function(:x))

Code like this won't break:

  dataset.filter(:number - 1 > 0)

Because :number - 1 yields a Sequel::SQL::ComplexExpression, which
defines #>.  Therefore, you can change the following:

  dataset.filter(:number > 1)

to:

  dataset.filter(:number - 1 > 0)

But I think using .sql_number is more clear.

In order to find existing uses of the above issues, here are a
couple commands you can run (from the root of your project):

  # Find :symbol[]
  egrep -nr ':['\''"]?[a-zA-Z_0-9]*['\''"]?\[' *
  # Find :symbol (<|>|<=|>=)
  egrep -nr '[^:]:['\''"]?[a-zA-Z_0-9]*['\''"]? *[<>]=?' *

All of the documentation was changed to reflect these new
conventions.  Note that if you are using ruby 1.8, you do not have
to make the change immediately.  However, if you ever plan to move to
ruby 1.9, it's better to make the change sooner than later.

In order to implement this change the Sequel::SQL::*Methods modules
were reorganized so that they don't include other modules.  This
means that if you were used to just including
Sequel::SQL::GenericExpressionMethods in a class to make it act like
a Sequel expression, you now need to include all modules separately.

Since Sequel::SQL::SpecificExpressionMethods and
Sequel::SQL::GenericExpressionMethods were modules whose sole purpose
was to include other modules, those modules have been removed.

Personally, I think making symbols more like strings was a bad
design decision, and hopefully one that will be removed in a later
version.  However, as ruby 1.9.1 is already at the release candidate
2 stage, I'm guessing it is here to stay.