Update CHANGELOG for tamas' before hook change.
The reasoning behind the change is that hooks should wrap
the existing functionality.  If there is already a before hook A
defined, adding a before hook B should have hook B run before hook
A, otherwise it is not wrapping properly.  This breaks backwards
compatibility, but is the correct way to handle things.