When using standard strings in the postgres adapter with the postgres-pr driver, use custom string escaping to prevent errors
postgres-pr is unmaintained AFAIK.  Instead of monkey patching to fix
the issue, change the escape_string method Sequel defines if
postgres-pr is used to check the force_standard_strings setting
and correctly escape if it is true (which it is by default).

postgres-pr has one other signficant problem, in that it raises
RuntimeErrors instead of PGErrors if there are errors in the SQL
syntax.  The fix isn't difficult, but as it would require modifying
an existing postgres-pr method, it's not one I'm comfortable making.