Fix literalization of SQL::Blobs in DataObjects and JDBC adapter's postgresql subadapters when ruby 1.9 is used
This code previously expected String#[] to return a Fixnum, but it
returns a String in ruby 1.9.  Change the code so that it works in
both ruby 1.8 and ruby 1.9.  Refactor the literalization code in
the postgres adapters and subadapters so that that the shared
PostgreSQL adapter has code that doesn't depend on the underlying
adapter.  The native postgres adapter now overrides #literal to
use the versions present in the adapters, as does the JDBC
postgres adapter for strings.