Massive changes: Identifier input/output converter methods, H2 JDBC subadapter, many fixes
This is a huge commit, and I'm too lazy to break it down into a
smaller one, since a lot of the parts are interdependent.

The biggest change is the addition of identifier input and output
methods.  The identifier_input_method is used when when literalizing
identifiers going into the database.  For example:

  DB[:table].insert(:column=>'value')

It will affect table and column, but not value since that is not an
identifier.  The default value is :upcase, since the SQL standard is
that unquoted identifiers are folded to uppercase, so you get:

  INSERT INTO "TABLE" ("COLUMN") VALUES ('value')

If you set the identifier_input_method to nil, which the MySQL,
PostgreSQL, and SQLite adapters do by default, you get:

  INSERT INTO "table" ("column") VALUES ('value')

The identifier_output_method is used when retrieving information
from the database.  Since SQL identifier are generally uppercase
(since that's what unquoted identifiers are folded to), and we
generally want to deal with them in lowercase, the default is
:downcase, so that you get lowercase identifiers even on a database
with uppercase identifiers.

Like most Sequel settings, this can be set globally, per database,
or per dataset:

  # Global
  Sequel.identifier_input_method = :downcase
  Sequel.identifier_output_method = :upcase
  # Per Database
  DB.identifier_input_method = :camelize
  DB.identifier_output_method = :underscore
  # Per Dataset
  ds = DB[:table]
  ds.identifier_input_method = :underscore
  ds.identifier_output_method = :reverse

I could have added a downcase_identifiers setting similar to the
upcase_identifiers setting, but that is confusing since they
operate differently.  The identifier input/output method naming
makes more sense, and gives you the flexibility to do something
crazy like :camelize and :underscore, which you may need to do
if you use CamelCase names in your database.

The upcase_identifiers setting now just sets the
identifier_input_method to :upcase (or nil if it is set to false).

The identifier_output_method is used inside of fetch_rows (actually,
a method that fetch_rows calls), and therefore it requires
changes to all adapters (since that's where the database specific
fetching happens).  As I only test 5 of the adapters and Sequel
supports 13, I need everyone that uses an adapter other than
MySQL, PostgreSQL, SQLite, JDBC, and DataObjects to test this
commit against their database and make sure I didn't break anything.
Unlike usual, not all of the changes I made to adapters I don't have
access to are small.  The most significant changes that I did not
test were to the firebird and informix adapters.

This commit adds an H2 subadapter for JDBC.  H2 is an embedded
database similar to SQLite, but it is easier for JRuby users to use
as it doesn't rely on any native code.  It passes all integration
tests with the exception of the schema tests, because it doesn't
support schema parsing yet.

This commit also adds a Database#tables method to the JDBC adapter,
and hopefully full schema parsing for JDBC will be coming soon.

This commit also changes the JDBC adapter so that Java specific
types such as Java::JavaSQL::Timestamp, Java::JavaSQL::Time,
and Java::JavaIo::BufferedReader are converted into normal ruby
types before being returned to the user.  This fixed a few bugs
in the intergation tests, and should make using the JDBC adapter
easier.

This commit fixes the literalization of Times and DateTimes when
using the JDBC MySQL subadapter.  It also fixes the use of
SQL::Blobs when using the JDBC PostgreSQL subadapter.

This commit refactors the MySQL native adapter so as to not
modify the Mysql::Result class.  This was necessary in order to
implement the identifier_output_method support.

This commit makes Database#quote_identifiers= have affect on
future usages of the schema modification methods such as
create_table, by removing the cached schema_utility_dataset.
The identifier input/output methods have the same effect.

This commit changes the default quoted_identifier method to
double any double quotes in the identifier, in addition to
surrounding the identifier in double quotes.

This commit changes the MySQL specs to only run the socket
specs if the native mysql adapter is being used, since it
uses the native mysql adapter to run those specs.

Due to implementation details,
Database.identifier_(input|output)_method returns "" if no
method should be used, and nil if it should fall back to the
adapter support.  You should not rely on this, though I don't plan
on changing it.

This commit breaks backwards compatibility in the following ways:

Using a database other than MySQL, PostgreSQL, or SQLite, if you
used to get uppercase symbols when retrieving records, you will
now probably get lowercase symbols.  To continue to receive
uppercase symbols:

  DB.identifier_output_method = nil

The Database#lowercase method in the DBI adapter was removed, as
it isn't needed since the default is to lowercase identifiers.