Clean up tamas's recent commit and update CHANGELOG
This is a potential breakage of backwards compatibility, but it's
unlikely that anyone would be relying on changed_columns being empty
in an after filter, or new? always returning false even the record
being saved was new.  With this commit, it's easier to code a complex
after_save that needs to differentiate between a freshly inserted
record and one that was updated (instead of using a separate
after_create and after_update filters), or have your after_update
filter act depending on contents of changed_columns (previously,
saving them to a separate instance variable in a before_update
filter was the best way).