Add validates_not_string validation, useful in conjunction with raise_on_typecast_failure = false
The RDoc gives a good description of the purpose of this:

Validates whether an attribute is not a string.  This is generally
useful in conjunction with raise_on_typecast_failure = false, where
you are passing in string values for non-string attributes (such as
numbers and dates).  If typecasting fails (invalid number or date),
the value of the attribute will be a string in an invalid format, and
if typecasting succeeds, the value will not be a string.