Add per-database type translation support for schema changes, translating ruby classes to database specific types
After many requests for per-database type translation, and my
repeatedly saying I'll accept a good patch, I went ahead and coded
the support myself.  Here's what the API looks like:

    DB.create_table(:cats) do
      String :a
      Integer :b
      Fixnum :c
      Bignum :d
      Float :e
      BigDecimal :f
      Date :g
      DateTime :h
      Time :i
      Numeric :j
      File :k
      TrueClass :l
      FalseClass :m
      column :n, Fixnum
      primary_key :o, :type=>String
      foreign_key :p, :f, :type=>Date
    end

Basically, anywhere you would use a symbol or string to specify a
type, you use the ruby class itself that you are trying to represent.
There is support in the MySQL, SQLite, and PostgreSQL adapters for
these mappings, and the defaults should be reasonable, with most
other databases probably only needing to override a few types.

This first form of column description was already accepted, so this
breaks compatibility.  In generally, most schema modification methods
would be specified in all lowercase, so this is not likely to affect
many people.  This also breaks compatibility by changing the meaning
of #Float, #Integer, and #String inside Sequel::SQL::Generator, since
they previously were Kernel private instance methods.   One of the
reasons I'm not as worried about compatibility is that since Float,
Integer, and String were already defined, using the above syntax
would not have worked well before, because String :a wouldn't create
a column (it would return "a"), and Integer :b and Float :e would
have raised errors.

This commit includes a big expansion to the integration type tests,
testing all of the generic types that Sequel supports.  Most
databases don't pass these yet.  SQLite fails on the File test
because it can't support "\0" values in strings.  PostgreSQL and
MySQL pass on ruby and ruby 1.9, but generally fail at least one test
when using JDBC or DataObjects.  The H2 JDBC subadapter fails on a
Timestamp column for reasons I don't quite understand, and also fails
the embedded "\0" values in strings test.

This commit updates the native SQLite adapter to convert float, real,
and double precision columns to Floats.

This commit adds support for conversion of Java::JavaSQL::Date to
Date and Java::JavaMath::BigDecimal to BigDecimal in the JDBC
adapter.