Add schema parsing support to the JDBC adapter
This commit adds support for the schema parser to the JDBC adapter
using the JDBC metadata methods.  It will be used when a JDBC
subadapter is used that doesn't have support for schema parsing in
a shared adapter.  Currently, that means that the following databases
should have schema parsing support when used with the JDBC adapter:
H2, MSSQL, and Oracle.  The schema parsing support was tested with the
H2 subadapter and it works for me.

This commit also includes an update to the H2 JDBC subadapter for
adding a primary key.  This doesn't appear to work yet (I get
"Timeout trying to lock table SYS" exception), but at least it
appears to be sending the correct SQL.  The H2 JDBC subadapter
doesn't appear to want to drop columns which are foreign key
references to other tables, but I consider that a database problem.
In total, this commit brings the number of broken integration tests
using the H2 JDBC subadapter from 12 to 4.

This commit also adds a private input_identifier method to Dataset,
similar to the existing output_identifier method.
Dataset#quote_identifier uses this now, and it is also used in
the JDBC schema parsing support.