Make Database#tables respect identifier_output_method
This updates Database#tables in the adapters to respect the
identifier_output_method.  I tested this on PostgreSQL and
SQLite, and the Firebird adapter maintainer says the
Firebird adapter change works.  This also modifies the
Oracle and JDBC adapters, which I haven't tested, but
hopefully they still work.  I have a fix for the MySQL
adapter, but I'm waiting to commit that until I can test it.

I added an integration test to test this change. It's fairly
simple but does a OK job.