Handle identifier_(in|out)put_method in adapters for tables and schema
This commit fixes some issues when identifier_input method or
identifier_output_method is used.  Because significant parts of
the Sequel adapters are written in the same abstractions that Sequel
makes available, changing the above methods would break things.

This commit attempts to fix breakage in the PostgreSQL, MySQL, JDBC,
and SQLite adapters.  For most adapters, this wasn't too difficult,
but it was quite a bit of work for the PostgreSQL adapter, for a
couple of reasons.  First, the default_schema in PostgreSQL is set
to :public instead of nil, and having the identifier_input_method
changed would change the schema it used by default, which broke a
lot of stuff.  I tryed changing the schema to be a LiteralString,
but that caused other issues.  Most of this commit is to deal with
these issues.

First, this removes the ability to load schema information all tables
at once.  This was only supported on PostgreSQL, and was only an
optimization.  Since it didn't work correctly, I removed it.  I also
made schema and table methods work if you change default_schema to
nil, and had to change default_schema itself so it didn't fallback to
the built in default if it was set to nil.

This commit also changes quote_identifier so that if you use a
LiteralString, it is returned without modification.  It also
adds a integration test for schema parsing when using
identifier_input_method and identifier_output_method, as well as some
helper methods for debug logging.