Add :clone association option, for making clones of existing associations
While Sequel makes it easy to get the dataset representing the
association via the association_dataset method, I recommend using a
separate association if you are doing the same thing with
association_dataset in multiple places.  This adds a shortcut to make
that easier, by allowing you to clone an existing association's
options.  This is especially helpful when dealing with legacy schemas
were you would have to specify many of the same options for each
association.

Here's an example use:

  many_to_many :deputies, :class=>:Employee, \
   :join_table=>:employeecurrentaudits, :left_key=>:currentauditid, \
   :right_key=>:employeeid, :order=>[:firstname, :lastname] do |ds|
    ds.filter(:active).filter(:capacity=>1)
  end
  many_to_many :project_managers, :clone=>:deputies do |ds|
    ds.filter(:active).filter(:capacity=>2)
  end
  many_to_many :team_leaders, :clone=>:deputies do |ds|
    ds.filter(:active).filter(:capacity=>3)
  end

Before, to do something like this, you would have had to specify
the :join_table, :left_key, :right_key, and :class options in each
association.  Also, though this example doesn't show it, the cloned
association's block is copied as well if the current association
doesn't use a block.  You can use the :block=>nil option to not
clone the previous block and not use a block for the current
association.