Add :conditions association option, for easier filtering of associated objects
This commit like the previous one doesn't allow anything that wasn't
possible before, but should make the API nicer.  Before, if you
wanted to filter associated records and cover all cases, you had to
specify a block and the :graph_conditions option.  Now, assuming
that your condition is expressible as a hash or array of two pairs,
you can just change :graph_conditions to :conditions and eliminate
the block.  You can change this:

  one_to_many :japanese_verses, :class=>:Verse,
   :graph_conditions=>{:languageid=>3} do |ds|
    ds.filter(:languageid=>3)
  end

to:

  one_to_many :japanese_verses, :class=>:Verse,\
   :conditions=>{:languageid=>3}