Make Dataset#select, #select_more, and #get take a block that yields a SQL::VirtualRow, similar to #filter
This is useful for select just like it is useful for filter, in that
it allows a nicer API in the case where you can't use the Symbol
shortcuts for methods (either because you are running on Ruby 1.9 or
because another library defines them).  So instead of:

  dataset.select(:prettify.sql_function(:name))

you can do:

  dataset.select{|o| o.prettify(:name)}

This commit also changes some of the dataset convenience methods to
use this new feature.