Make PostgreSQL and Firebird adapters send inserts to the master server instead of the slave server by default
The PostgreSQL (if it is connecting to a 8.2 or greater database) and
Firebird adapters use INSERT RETURNING to return data directly from
the INSERT query.  Because of that, the INSERT query is treated as a
SELECT by Sequel.  Since it is treated as a select, the request went
to the slave servers instead of the master.  Fix this by setting the
server to use to the master (:default) server unless it is already
set to use a different server.

To easily implement this, add Dataset#default_server_opts private
method.

This comment also removes pointless execute_insert definitions some
adapters, since they just duplicated the now default definition.