Make Dataset#literal call methods like #literal_string, override less in adapters, for a 20-25% performance increase
Profiling shows Dataset#literal can take around 10% total time, so
this is probably only a 2% performance increase in general.

Basically, this changes how the adapters override the default
literalization.  Instead of overriding literal and using a case
statement and super in the else clause, they override literal_*
methods, so the default Dataset#literal is always called.  In
addition to being faster the code is noticeably cleaner as well.

During this change I finally decided to change the Sequel default
date and timestamp formats from the SQL standard (e.g.
"TIMESTAMP 'YYYY-MM-DD HH:MM:SS'") to the defacto ISO8601 standard used
by most of the adatpers.  I'm pretty sure this won't break existing
code, but testing on most adapters would be nice, especially on
the Firebird and ODBC adapters.

This commit also moves the unsupported.rb and stored_procedures.rb
from lib/sequel_core/dataset to lib/sequel_core/adapters/utils, since
they are only used by certain adapters.  One of the results of this
is that Sequel::Dataset::UnsupportedExceptIntersect(All)? is not
guaranteed to be defined.

From my profiling, it appears that another 10% increase to
Dataset#literal could be achieved by adding a #to_sequel_literal
method to all objects, and calling that instead of using a case
statement.  At this point, that's not a road I want to go down.