Add an optimization to Model.[] for simple models, with a 30-40% performance increase
After some profiling I discovered that Model.[] can spend almost half
of it's time in producing the SQL to send to the database.  This is
because it goes through the entire SQL creation process, which is
very flexible, but not optimized for performance.  If it's possible
to use static SQL instead of going through the usual SQL generation
process, there would be a dramatic performance increase.

This adds a Database#find_by_table_column_value method that uses
a given literal table and column, and a value argument, and returns
the first matching row (with all columns) for that table.  The value
argument is literalized, but the table and column are used as is,
so you cannot use untrusted input for those arguments.

If a model is a optimizable (single primary key, single table),
find_by_table_column_value is used instead of the usual lookup using
primary_key_hash.  Unfortunately, it wasn't easy to determine which
models are optimizable.  Checking all of the dataset options for the
model dataset is too likely to break in future versions, even if I
did get it correct in this version.  I took the conservative approach
and made it so that the model is optimizable if a symbol is passed to
set_dataset. So the following models will use the optimization:

  class Foo < Sequel::Model
    # Using the implicit table is optimized
  end
  class Foo < Sequel::Model(:foos)
    # Since you provide a Symbol to Model(), it is optimized
  end
  class Foo < Sequel::Model
   set_dataset :foos
   # Symbol provided to set_dataset is optimized
  end

But the following otherwise equivalent models are not optimized:

  class Foo < Sequel::Model(Model.db[:foos])
  end
  class Foo < Sequel::Model
   set_dataset db[:foos]
  end

To make this as fast as possible, two class instance variables are
added to model: @simple_pk and @simple_table.  They store the
already literalized table and primary key column names, which are
passed to find_by_table_column_value.

This commit also adds some specs for the previous optimization
commit, to bring code coverage back to 100%.