Allow use of database independent types when casting
Previously, types were used as-is when casting.  Now, casting uses
the same type casting code as the schema methods, so you can specify
ruby classes and have them changed to database specific types.

This changes the object used for casts from SQL::IrregularFunction
SQL::Cast.  SQL::IrregularFunction is now just used for handling
the SQL EXTRACT function (though it may be used for other similar
functions in the future).

This adds an integration test for the feature (in addition to the
usual specs).  It fails on SQLite because of the way SQLite handles
casting (returning strings even when you are casting to integer).