Remove Database#find_by_table_column_value, add Dataset#with_sql, for optimization in Model.[]
I shouldn't have added a database method for this, so back it out.
The dataset method makes more sense, and I've been wanting to add it
for a while.  Dataset#static_sql was already taken, so I chose
Dataset#with_sql.  Add the usual dataset mutation method and model
dataset methods, and switch Model.[] to use with_sql.

This new code allows the optimized version to work with models with
serialized columns, plus models that use datasets with nonstandard
row_procs.

While here, fix Model.set_dataset so that it removes the simple_table
if a dataset is given.