Allow String.lit to take arguments and return a SQL::PlaceholderLiteralString
This useful when you want an SQL fragment to be literalized, but also
want it to contain other values.  Before, this was hard to do without
calling Dataset#literal.  Now, it's much easier to deal with:

  column = :user
  pattern = params[:pattern]
  dataset.select{|o| o.substring('? from ?'.lit(column, pattern))}

This commit also changes the internals so that String#lit is never
called, so Sequel itself won't break if another library defines it.