Remove SQL::IrregularFunction, no longer used internally
The only user of this class left was from
ComplexExpressionMethods#extract, which was changed to use a
regular Function plus a PlaceholderLiteralString.
SQL::IrregularFunction was always a bad hack, and the function
plus placeholder string method is much neater and better all
around.

This also fixes another String#lit call in the shared SQLite adapter.