Allow virtual row blocks to be instance_evaled, add Sequel.virtual_row_instance_eval=
This commit allows the user to use instance_eval for virtual row
blocks that do not have arguments.  The user needs to first change
the setting:

  Sequel.virtual_row_instance_eval = true

If this is not set and a virtual row block without an argument is
used, a deprecation message is printed.  VirtualRow blocks are used
by filter, order, select, and methods that call those methods with
blocks (quite a few, see command line below).  It allows you to
change code that looks like this:

  dataset.filter{|o| (o.number > 10) & (o.name > 'M')}

to:

  dataset.filter{(number > 10) & (name > 'M')}

When instance_eval is used, only local variables are available
to the block.  Any calls to instance methods will be interpreted
as calling VirtualRow#method_missing, which generates identifiers
or functions.  With this commit and when virtual_row_instance_eval
is enabled, the following type of code will break:

  # amount is a instance method
  dataset.filter{:number + amount > 0}

Just like this example, the only type of code that should break is
when you used a virtual row block when it wasn't necessary (since
you weren't using the VirtualRow argument).  I don't expect that to
be a common practice, but as I didn't want to break existing code,
I choose to deprecate it, and remove it in a future version.

In a future version of Sequel, the usage of instance_eval for
virtual row blocks without block arguments will not be optional, it
will be the only way Sequel works.

In order to start printing deprecation messages,
sequel_core/deprecated is once again required, though I changed the
implementation a little.

In order to find places in your code where this deprecation may
affect you, try the following command line:

  egrep -nr '[^A-Za-z0-9_](filter|select|select_more|order|order_more|get|where|having|from|first|and|or|exclude|find|subset|constraint|check)( *(\([^)]*\) *)?){ *[^|]' *

It's not perfect (it misses cases where you use a block and regular
arguments which contain parantheses), but it should catch most cases.