Deprecate Sequel.open, Sequel.use_parse_tree=?, and the upcase_identifier methods
Sequel.open is just an alias of Sequel.connect, and I think connect
is a better method name.

The use_parse_tree methods haven't had an effect since before
Sequel 2.2.

The upcase_identifier methods have been superceded by the
identifier_input_method methods.

This also deprecates the :upcase_identifiers database option.

This includes a slight refactor to Deprecation.deprecate, to reduce
verbosity.  It also adds a deprec_specify RSpec helper, to silence
the printing of deprecation methods for certain specs.  The reason
I did this is so if I deprecate something that is used in another
spec, I want to see the deprecation message (because it probably
needs to be fixed).  However, if I know I'm using a deprecated
method in a spec, I probably don't want to see the message.  This
also has the advantage of making the specs easy to clean up after
the deprecated methods are removed.