Deprecate Enumerable#send_each and, on ruby 1.8, Hash#key
This also changes Sequel so that it doesn't add Hash#key if it is
already present.  This way, if you are using another library that
adds Hash#key, Sequel won't override it with a method that prints
a deprecation message.

Enumerable#send_each wasn't used at all by Sequel, and Hash#key
was only used in the shared SQLite adapter.  I refactored that
by changing some hash constants to array constants.