Deprecate passing options to Dataset#each, #all, #single_record, #single_value, #sql, #select_sql, #update, #update_sql, #delete, #delete_sql, #exists
Allowing some of these methods to take options has caused numerous
bugs. I think the serious ones have been fixed, but I'm guessing
there are still bugs.  Forcing users to clone datasets with options
before calling these methods ensures that the dataset's options and
the options that each is using are the same.

The other reason that allowing options to be passed is bad is that it
exposes more of the internal implementation than it should.  We
should be encouraging the use of:

  from(:table).all

instead of:

  all(:from=>[:table])