Deprecate Dataset#uniq
This deprecation is more serious than some others, as this deprecates
the original method name in favor of the original alias.  I think
distinct is the better name, and I don't like the fact that
Dataset#uniq is too similar to Array#uniq.

This commit also changes the MySQL and Oracle adapters to raise an
error when Dataset#distinct is called, instead of waiting until the
SQL is generated.