Deprecate Model.is and Model.is_a, use Model.plugin for plugins
This is probably more controversial than the other deprecations,
since the replacement method is new.  I feel that is and is_a are
poor enough choices for method names that they deserve to be
deprecated.  Making the plugin-adding method name read more like
standard English causes more problems then it solves.  For example,
take the sequel_polymorphic plugin.  You add it with:

  is :polymorphic

Which is misleading, as the model itself it not polymorphic, it
just gives you the ability to use polymorphic associations.  The
problem is that:

  is :polymorphic_associations
  is_a :polymorphic_associations

Both don't read like standard English, which is probably why the
misleading name is chosen.  The use of a more intuitive, more
descriptive, but less grammatical method name should improve code
readability in the long run.