Deprecate Model#save!, use Model.save(:validate=>false)
Model#save! is a bad idea.  When I first took over as maintainer,
I thought it meant raise an error if the save was unsuccessful,
and it caused bugs in earlier versions of the one_to_many
association add/remove methods.  This changes it to use
something more verbose but a lot more descriptive.