Speed up the calling of most column accessor methods, and reduce memory overhead of creating them
The previous implementation of def_column_accessor leaked a lot of
instance method name string objects upon method creation, and always
used define_method with a block to create methods, which is slower
than a normal method call due to the closure.  This new
implementation fixes both issues.  It uses module_eval with a string
for method creation when it is possible, and define_method with a
block inside a separate method when it is not possible, so there is
only one local variable used by the closure.

tmm1 reports this saves almost half a MB of memory in his app.