Deprecate String#to_date, #to_datetime, #to_time, and #to_sequel_time, use require 'sequel/extensions/string_date_time' if you want them
This commit deprecates the date/time string conversion methods, and
adds string_to_* methods to the Sequel module.  It modifies Sequel
to use the module methods.

This required some refactoring of the MySQL adapter, which was
designed to only call methods of the object.  I modified it
to act more like the PostgreSQL adapter, where it takes procs.
I went a step further and refactored both the MySQL and PostgreSQL
adapters to reduce the number of proc objects used.