Deprecate model validation class methods, just override Model#validate or use Model.plugin :validation_class_methods
The validation code was always more complex than it needed to be,
which isn't surprising since it tried to mimic ActiveRecord.  Using
instance methods is simpler and probably performs better too:

  def validate
    errors[:finish_date] << "before start date" if finish_date < start_date
  end

as opposed to:

  validates_each(:finish_date) do |o,a,v|
    o.errors[a] << "before start date" if v < o.start_date
  end

There currently aren't any helper methods for instance level
validations.  I'm not sure if Sequel will be providing those or not,
as that is something easily solved by a plugin.

Refactor both the deprecated validations and the new plugin's handling
of the :if option, for a minor speedup.

Reorder the sequel_model requires, and make base.rb not depend on
associations.rb.  Require hooks.rb and deprecated_validations.rb from
deprecated.rb, so it will be easier to disable them.

Move #errors, #validate, and #valid? to Model::InstanceMethods.